Show chameleon sites instead of regions
This adds a new way to render the top navigation. Instead of showing the
regions as dropdowns, instead show a list of Chameleon sites, fetched
from the reference API.

To opt-in to this functionality an operator can define:

  CHAMELEON_MULTISITE_SUPPORT = True

Additionally, the site ID needs to be specified. This needs to map to an
ID of a site in the reference API:

  CHAMELEON_SITE_ID = "tacc"