Support redirect 'next' param in WebSSO
If a user logs in via WebSSO, the 'next' parameter is not respected, and
so the user ultimately must re-navigate to their target destination once
they complete their login. This fixes that by including the 'next'
parameter on the Keystone WebSSO start request. Keystone will ignore
this flag. However, we can then read it from the referer when Horizon is
handling its final part of the login flow.