Add multi-site support
This adds a new way to render the top navigation. Instead of showing the
regions as dropdowns, instead show a list of Chameleon sites, fetched
from the reference API. This allows us to further obscure the
multi-region capabilities of Chameleon, with the aim of ultimately
removing them.

To opt-in to this functionality an operator can define:

  CHAMELEON_MULTISITE_SUPPORT = True

Additionally, the site ID needs to be specified. This needs to map to an
ID of a site in the reference API:

  CHAMELEON_SITE_ID = "tacc"

Finally, to disable the existing multi-region dropdown:

  OPENSTACK_KEYSTONE_MULTIREGION_SUPPORT = False

This only affects whether the UI is shown and doesn't modify any other
functionality/capabilities.