Use global login opt-in cookie for federation auth
This updates the login method detection to use the global opt-in cookie
set by the Chameleon user portal. Because reading a cookie requires
access to the request object, a few function signatures had to change to
make this possible.

This works by telling Horizon that there is no default redirect URL if
this cookie is set; that default URL was what was providing the
automatic redirect to Portal for the legacy SSO.