Support forcing back to old login experience
This adds support for a GET param ?old_login_experience that lets you
override the cookie to do a one-off login using the old way.

It also adds an /auth/force-password-login path that redirects to login
with this param. This endpoint is rendered in the Keycloak IdP to allow
users to attempt to login w/o federation.