Drop Django 2.2 support
Django 2.2 support was dropped. Django 2.2 ends its extended support
in April 2022[1]. Considering this, horizon dropped Django 2.2 support
and use Django 3.2 as default version of Django.

It also updates the python version from py38 to py39 for
"horizon-non-primary-django-jobs" and
"horizon-tox-python3-django" project template because py36
and py39 are now runtime for the Yoga cycle[2]. So I have updated it to
py39.

[1] https://www.djangoproject.com/download/
[2] https://review.opendev.org/c/openstack/governance/+/820195

Change-Id: I8886f12f7e51b721ac5123070086d5dcfefe4e72