local_settings.py: Disable django.template DEBUG messages
VariableDoesNotExist error in DEBUG messages from django.template are
super noisy. It happens even for cases where variables are not defined
like "default" filter. It also happens with Django built-in templates.
Previously we tried to fix VariableDoesNotExist errors in our templates,
but it tunrs out there is no convenient way to prevent it.

We agreed to disable django.template DEBUG messages by default
in the horizon weekly meeting [1].

Closes-Bug: #1720893

[1] https://meetings.opendev.org/meetings/horizon/2022/horizon.2022-02-23-15.00.log.html#l-105

Change-Id: Ife9a91c7cd050b30d7f1eca8861a0d6629a23827