Make integration tests stop on first failure
This should save us some time and decrease the server load. If a single
test failed, we already know the patch is bad (or the tests need to be
rerun) so it makes no sense to execute the rest of the tests on the
gate.

Change-Id: Ic56d5e192487d444afd0cdf0bce5e1a691ba087c
(cherry picked from commit 4eb24797e53a39ec42c70768c160a838400aeadb)