Pin tox<4 for stable branches testing
horizon npm jobs start failing due to some recent changes in tox4.
As discussed on open-discuss ML, it is decided to keep testing the
stable branch with the tox version that they release with and not
with the tox 4. This patch set the 'ensure_tox_version' to use 4,
which will tell zuul to install tox<4 if it's not present.
For more info. please refer [1].

[1] https://lists.openstack.org/pipermail/openstack-discuss/2023-January/031668.html

Change-Id: I3652d87d817030bcd3855d26bfa41b8b7061d08e