Remove unused tabs (#20)
This patch removes several tabs from the main dashboard menu and the instance creation modal dialogue. These tabs were either completely unsupported on baremetal sites or simply unused. 

The tabs removed:

* Security Groups
    * Removed from dashboard and instance creation dialogue
* Flavors
    * Removed from instance creation dialogue
* Server Groups
    * Removed from dashboard and instance creation dialogue

Tabs removed from the main dashboard simply don't execute the code which registers them to that dashboard. 

The instance creation modal dialogue unfortunately only exists in the front end, so the code which registers the undesired tabs is simply commented out. 

The "baremetal" flavor is now picked automatically on the backend when a new instance is created.

Since the instance creation modal dialogue only exists in the front-end, it was not obvious if it was possible to transfer the context of this new setting to the dialogue to allow for conditional registration of the undesired tabs. Instead, the tabs are simply commented out. Because of this, these particular commits should not be deployed on KVM.