fix quota display when only one quota is shown
When the displayed quotas are overridden to only show one chart,
(such as showing instances, but not vcpu/memory under compute)
the "row" html tag was not closed.

While this could be fixed by removing the check for "forloop.first" in
the closing bounds check, a more robust fix seems to be moving the row
open/close tags outside of the loop entirely.

Those checks were only in place to generate a new row every 6 columns,
but that is now handled by the bootstrap column spec instead, so we can
unconitionally generate a row for each section, and let bootstrap handle
rendering/wrapping of the columns.