fix error handling for default flavor
our logic to set the default flavor if only one is defined had the
potential to bypass the flavor validation logic. to resolve this, it was
moved outwards from the "build" to "update" method, and now checks the
"enabled" property, which is set after validation.

Additionally, logic was adjusted for the validation method, which would
raise errors when allocation flavor was set programatically, as the form
was not yet initialized.