BAREMETAL HACK: hide non-baremetal flavor info
the baremetal flavor doesn't make use of vcpus, ram, disk, and so on.
This is difficult to hide for ONLY a baremetal flavor, so instead just
rip the columns and charts out.

TODO: This will need to be reverted if VMs are deployed on the same
site as baremetal, or if we ever want to deploy this fork of horizon for
KVM.

A potential alternative is to apply this change via horizon override.
see: https://docs.openstack.org/horizon/latest/configuration/customizing.html#horizon-customization-module-overrides