instance-launch form: fix visibility of "advanced"
if a tab sequentiially "after" a tab marked "advanced" was selected,
all advanced tabs became visisble. Removing this code fixes that
behavior. Now, if an advanced tab is selected, then "show advanced" is
toggled off, the hidden tab will remain selected.

Since this matches the documented behavior in the commit that introduced
this, I propose that it's ok.